+++ CHANGELOG.md
@@ -5,6 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+
+- New `ContextMenu` component to show custom context menus on right-click.
+- `UIInstanceManager.onActive` and `UIInstanceManager.onInactive` lifecycle events for UI variant status changes.
+- Default `PlayerContextMenu` to access player information directly from the UI.
+- `PlayerInsightsPanel` to show detailed player insights and diagnostics.
+
 ## [4.15.1] - 2026-06-11
 
 ### Fixed